+++ NEWS.md
@@ -4,6 +4,8 @@
 `httr::progress`.
 - Added `withProgressWaitress`, `setProgressWaitress`, and
 `incProgressWairess` analogous to the same functions in shiny.
+- Added `withProgressAttendant`,
+`incProgressAttendant`, and `setProgressAttendant`.
 
 # waiter 0.2.3