+++ docs/changelog.md
@@ -8,6 +8,8 @@ List of changes made to waiter.
 `httr::progress`.
 - Added `withProgressWaitress`, `setProgressWaitress`, and
 `incProgressWairess` analogous to the same functions in shiny.
+- Added `withProgressAttendant`,
+`incProgressAttendant`, and `setProgressAttendant`.
 
 ## waiter 0.2.3